+++ lms/views/decorators/__init__.py
@@ -19,9 +19,5 @@ For documentation see:
 https://docs.pylonsproject.org/projects/pyramid/en/latest/narr/viewconfig.html#non-predicate-arguments
 """
 from lms.views.decorators.lis_result_sourcedid import upsert_lis_result_sourcedid
-from lms.views.decorators.reports import report_lti_launch
 
-__all__ = (
-    "report_lti_launch",
-    "upsert_lis_result_sourcedid",
-)
+__all__ = ("upsert_lis_result_sourcedid",)
